# docs/sprint-1/openapi.yaml
openapi: 3.1.0
info:
  title: SmartSweep API
  version: 0.1.0
paths:
...
  /api/v1/complaints:
    post:
      tags:
      - Complaints
      summary: Create Complaint
      operationId: create_complaint_api_v1_complaints_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplaintSubmit'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Complaints
      summary: List Complaints
      operationId: list_complaints_api_v1_complaints_get
      security:
      - HTTPBearer: []
      parameters:
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: ward_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Ward Id
      - name: complaint_type
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/ComplaintType'
          - type: 'null'
          title: Complaint Type
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/ComplaintCategory'
          - type: 'null'
          title: Category
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_ComplaintRead_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/complaints/high-risk:
    get:
      tags:
      - Complaints
      summary: High Risk Complaints
      operationId: high_risk_complaints_api_v1_complaints_high_risk_get
      security:
      - HTTPBearer: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_ComplaintRead_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/complaints/upload-photo:
    post:
      tags:
      - Complaints
      summary: Upload Complaint Photo
      operationId: upload_complaint_photo_api_v1_complaints_upload_photo_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_complaint_photo_api_v1_complaints_upload_photo_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Upload Complaint Photo Api V1 Complaints Upload Photo
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/complaints/{complaint_id}/photo:
    post:
      tags:
      - Complaints
      summary: Attach Complaint Photo
      description: Validate and attach an evidence photo to an existing complaint.
      operationId: attach_complaint_photo_api_v1_complaints__complaint_id__photo_post
      security:
      - HTTPBearer: []
      parameters:
      - name: complaint_id
        in: path
        required: true
        schema:
          type: integer
          title: Complaint Id
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_attach_complaint_photo_api_v1_complaints__complaint_id__photo_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/complaints/duplicate-check:
    post:
      tags:
      - Complaints
      summary: Duplicate Check
      description: Return advisory matches for a complaint draft before submission.
      operationId: duplicate_check_api_v1_complaints_duplicate_check_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DuplicateCheckRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  additionalProperties: true
                  type: object
                type: array
                title: Response Duplicate Check Api V1 Complaints Duplicate Check
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/complaints/{complaint_id}:
    get:
      tags:
      - Complaints
      summary: Get Complaint
      operationId: get_complaint_api_v1_complaints__complaint_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: complaint_id
        in: path
        required: true
        schema:
          type: integer
          title: Complaint Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Complaints
      summary: Update Complaint
      operationId: update_complaint_api_v1_complaints__complaint_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: complaint_id
        in: path
        required: true
        schema:
          type: integer
          title: Complaint Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplaintUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/complaints/{complaint_id}/status:
    patch:
      tags:
      - Complaints
      summary: Change Complaint Status
      operationId: change_complaint_status_api_v1_complaints__complaint_id__status_patch
      security:
      - HTTPBearer: []
      parameters:
      - name: complaint_id
        in: path
        required: true
        schema:
          type: integer
          title: Complaint Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Body_change_complaint_status_api_v1_complaints__complaint_id__status_patch'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/complaints/{complaint_id}/verify:
    patch:
      tags:
      - Complaints
      summary: Verify Complaint
      operationId: verify_complaint_api_v1_complaints__complaint_id__verify_patch
      security:
      - HTTPBearer: []
      parameters:
      - name: complaint_id
        in: path
        required: true
        schema:
          type: integer
          title: Complaint Id
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/ComplaintVerify'
              - type: 'null'
              title: Verify In
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/complaints/{complaint_id}/close:
    patch:
      tags:
      - Complaints
      summary: Close Complaint
      operationId: close_complaint_api_v1_complaints__complaint_id__close_patch
      security:
      - HTTPBearer: []
      parameters:
      - name: complaint_id
        in: path
        required: true
        schema:
          type: integer
          title: Complaint Id
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/ComplaintClose'
              - type: 'null'
              title: Close In
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/complaints/{complaint_id}/cancel:
    post:
      tags:
      - Complaints
      summary: Cancel Complaint
      operationId: cancel_complaint_api_v1_complaints__complaint_id__cancel_post
      security:
      - HTTPBearer: []
      parameters:
      - name: complaint_id
        in: path
        required: true
        schema:
          type: integer
          title: Complaint Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/complaints/{complaint_id}/history:
    get:
      tags:
      - Complaints
      summary: Complaint History
      operationId: complaint_history_api_v1_complaints__complaint_id__history_get
      security:
      - HTTPBearer: []
      parameters:
      - name: complaint_id
        in: path
        required: true
        schema:
          type: integer
          title: Complaint Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComplaintStatusHistoryRead'
                title: Response Complaint History Api V1 Complaints  Complaint Id  History
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/complaints/{complaint_id}/feedback:
    post:
      tags:
      - Complaints
      summary: Submit Complaint Feedback
      operationId: submit_complaint_feedback_api_v1_complaints__complaint_id__feedback_post
      security:
      - HTTPBearer: []
      parameters:
      - name: complaint_id
        in: path
        required: true
        schema:
          type: integer
          title: Complaint Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Complaints
      summary: Get Complaint Feedback
      operationId: get_complaint_feedback_api_v1_complaints__complaint_id__feedback_get
      security:
      - HTTPBearer: []
      parameters:
      - name: complaint_id
        in: path
        required: true
        schema:
          type: integer
          title: Complaint Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/complaints/{complaint_id}/classify:
    post:
      tags:
      - Complaints
      summary: Classify Complaint
      description: Classify a complaint's hazard category via Claude, falling back
        to keyword matching.
      operationId: classify_complaint_api_v1_complaints__complaint_id__classify_post
      security:
      - HTTPBearer: []
      parameters:
      - name: complaint_id
        in: path
        required: true
        schema:
          type: integer
          title: Complaint Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintClassifyRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/complaints/{complaint_id}/duplicates:
    get:
      tags:
      - Complaints
      summary: Get Duplicates
      operationId: get_duplicates_api_v1_complaints__complaint_id__duplicates_get
      security:
      - HTTPBearer: []
      parameters:
      - name: complaint_id
        in: path
        required: true
        schema:
          type: integer
          title: Complaint Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response Get Duplicates Api V1 Complaints  Complaint Id  Duplicates
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    Body_attach_complaint_photo_api_v1_complaints__complaint_id__photo_post:
      properties:
        photo:
          type: string
          format: binary
          title: Photo
      type: object
      required:
      - photo
      title: Body_attach_complaint_photo_api_v1_complaints__complaint_id__photo_post
    Body_change_complaint_status_api_v1_complaints__complaint_id__status_patch:
      properties:
        status_value:
          $ref: '#/components/schemas/ComplaintStatus'
      type: object
      required:
      - status_value
      title: Body_change_complaint_status_api_v1_complaints__complaint_id__status_patch
    Body_upload_complaint_photo_api_v1_complaints_upload_photo_post:
      properties:
        photo:
          type: string
          format: binary
          title: Photo
      type: object
      required:
      - photo
      title: Body_upload_complaint_photo_api_v1_complaints_upload_photo_post
...
    ComplaintCategory:
      type: string
      enum:
      - None
      - Foul Smell
      - Overflowing Bin
      - Mosquito Breeding
      - Risk to Children
      title: ComplaintCategory
      description: Hazard classification. Mirrors the ORM enum (US-28/29/30).
...
    ComplaintRead:
      properties:
        title:
          type: string
          maxLength: 255
          minLength: 1
          title: Title
        description:
          type: string
          minLength: 1
          title: Description
        ward_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Ward Id
        category:
          anyOf:
          - $ref: '#/components/schemas/ComplaintCategory'
          - type: 'null'
        complaint_type:
          anyOf:
          - $ref: '#/components/schemas/ComplaintType'
          - type: 'null'
        priority:
          anyOf:
          - type: string
            maxLength: 50
          - type: 'null'
          title: Priority
        address:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Address
        latitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Latitude
        longitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Longitude
        photo_url:
          anyOf:
          - type: string
            maxLength: 10000000
          - type: 'null'
          title: Photo Url
        tags:
          items:
            type: string
          type: array
          title: Tags
        id:
          type: integer
          title: Id
        reported_by_user_id:
          type: integer
          title: Reported By User Id
        status:
          $ref: '#/components/schemas/ComplaintStatus'
        resolved_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Resolved At
        cancelled_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Cancelled At
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - title
      - description
      - id
      - reported_by_user_id
      - status
      - created_at
      - updated_at
      title: ComplaintRead
      description: Complaint response schema.
...
    ComplaintSubmit:
      properties:
        location:
          type: string
          maxLength: 255
          minLength: 1
          title: Location
        description:
          type: string
          minLength: 1
          title: Description
        hazard:
          anyOf:
          - $ref: '#/components/schemas/ComplaintCategory'
          - type: 'null'
        complaint_type:
          anyOf:
          - $ref: '#/components/schemas/ComplaintType'
          - type: 'null'
        photo:
          anyOf:
          - type: string
            maxLength: 10000000
          - type: 'null'
          title: Photo
        coords:
          anyOf:
          - additionalProperties:
              type: number
            type: object
          - type: 'null'
          title: Coords
        ward_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Ward Id
      type: object
      required:
      - location
      - description
      title: ComplaintSubmit
      description: Payload coming directly from the citizen report form.
    ComplaintType:
      type: string
      enum:
      - overflow
      - delay
      - extra_collection
      title: ComplaintType
      description: 'Community-facing complaint category (S2-A18).


        Separate from ``category`` (the hazard classification field). Pydantic

        validates any value outside this set as a 422 automatically.'
...
    ComplaintVerify:
      properties:
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
          description: Review notes by admin
      type: object
      title: ComplaintVerify
      description: Payload for verifying a complaint (Admin review).
...
    DuplicateCheckRequest:
      properties:
        location:
          type: string
          maxLength: 255
          minLength: 1
          title: Location
        description:
          type: string
          minLength: 1
          title: Description
        coords:
          anyOf:
          - additionalProperties:
              type: number
            type: object
          - type: 'null'
          title: Coords
        ward_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Ward Id
      type: object
      required:
      - location
      - description
      title: DuplicateCheckRequest
      description: Draft complaint fields used for an advisory duplicate scan.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Page_ComplaintRead_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ComplaintRead'
          type: array
          title: Items
        meta:
          $ref: '#/components/schemas/PageMeta'
      type: object
      required:
      - items
      - meta
      title: Page[ComplaintRead]
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer